eagle/requirements.txt
black==21.10b0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4